refactor(android): move google maps to library
