+++ actor-sdk/sdk-core-android/android-app/build.gradle
@@ -38,5 +38,6 @@ dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile project(':actor-sdk:sdk-core-android:android-sdk')
     compile project(':actor-sdk:sdk-core-android:android-google-push')
+    compile project(':actor-sdk:sdk-core-android:android-google-maps')
 
 }
