+++ actor-sdk/sdk-core-android/android-google-maps/src/main/AndroidManifest.xml
@@ -1,0 +1,12 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="im.actor.maps.google">
+
+
+    <application
+        android:allowBackup="true"
+        android:supportsRtl="true">
+
+
+    </application>
+
+</manifest>
