+++ actor-sdk/sdk-core-android/android-google-maps/src/main/java/im/actor/map/PlaceFetchingTask.java
@@ -1,4 +1,4 @@
-package im.actor.sdk.controllers.pickers.map;
+package im.actor.map;
 
 import android.os.AsyncTask;
 import android.util.Log;
@@ -14,6 +14,8 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLEncoder;
 import java.util.ArrayList;
+
+import im.actor.map.MapItem;
 
 /**
  * Created by kiolt_000 on 22/09/2014.
