+++ actor-sdk/sdk-core-android/android-google-maps/src/main/res/layout/picker_activity_map_picker.xml
