+++ actor-sdk/sdk-core-android/android-google-maps/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="app_name">android-google-maps</string>
+
+</resources>
