+++ actor-sdk/sdk-core-android/android-sdk/src/main/java/im/actor/sdk/controllers/pickers/Intents.java
@@ -4,7 +4,7 @@ import android.content.Context;
 import android.content.Intent;
 
 import im.actor.sdk.controllers.pickers.file.FilePickerActivity;
-import im.actor.sdk.controllers.pickers.map.MapPickerActivity;
+
 
 /**
  * Created by ex3ndr on 14.10.14.
@@ -16,7 +16,8 @@ public class Intents {
 
 
     public static Intent pickLocation(Context context) {
-        return new Intent(context, MapPickerActivity.class);
+        Intent intent = new Intent("im.actor.pickLocation");
+        return intent;
     }
