+++ settings.gradle
@@ -25,6 +25,9 @@ include ":actor-sdk:sdk-core-android:android-sdk"
 //Google push for Android SDk
 include ":actor-sdk:sdk-core-android:android-google-push"
 
+//Google maps for Android SDk
+include ":actor-sdk:sdk-core-android:android-google-maps"
+
 //
 // SDK iOS
 //
